+++ crates/objc2/src/macros/mod.rs
@@ -696,7 +696,8 @@ macro_rules! __class_inner {
 ///
 /// The first expression, know as the "receiver", can be any type that
 /// implements [`MessageReceiver`], like a reference or a pointer to an
-/// object, or even a reference to an [`rc::Id`] containing an object.
+/// object. Additionally, it can even be a reference to an [`rc::Id`]
+/// containing an object.
 ///
 /// The expression can be wrapped in `super`, with an optional superclass
 /// as the second argument. If no specific superclass is specified, the
@@ -708,10 +709,10 @@ macro_rules! __class_inner {
 /// If the last argument is the special marker `_`, the macro will return a
 /// `Result<(), Id<E>>`, see below.
 ///
-/// This macro translates into a call to [`sel!`], and afterwards a fully
-/// qualified call to [`MessageReceiver::send_message`]. Note that this means
-/// that auto-dereferencing of the receiver is not supported, and that the
-/// receiver is consumed. You may encounter a little trouble with `&mut`
+/// This macro roughly translates into a call to [`sel!`], and afterwards a
+/// fully qualified call to [`MessageReceiver::send_message`]. Note that this
+/// means that auto-dereferencing of the receiver is not supported, and that
+/// the receiver is consumed. You may encounter a little trouble with `&mut`
 /// references, try refactoring into a separate method or reborrowing the
 /// reference.
 ///
@@ -952,11 +953,11 @@ macro_rules! __class_inner {
 macro_rules! msg_send {
     [super($obj:expr), $($selector_and_arguments:tt)+] => {
         $crate::__msg_send_parse! {
-            (__send_super_message_static_error)
+            (send_super_message_static_error)
             ()
             ()
             ($($selector_and_arguments)+)
-            (__send_super_message_static)
+            (send_super_message_static)
 
             ($crate::__msg_send_helper)
             ($obj)
@@ -964,7 +965,7 @@ macro_rules! msg_send {
     };
     [super($obj:expr, $superclass:expr), $($selector_and_arguments:tt)+] => {
         $crate::__msg_send_parse! {
-            (__send_super_message_error)
+            (send_super_message_error)
             ()
             ()
             ($($selector_and_arguments)+)
@@ -976,7 +977,7 @@ macro_rules! msg_send {
     };
     [$obj:expr, $($selector_and_arguments:tt)+] => {
         $crate::__msg_send_parse! {
-            (__send_message_error)
+            (send_message_error)
             ()
             ()
             ($($selector_and_arguments)+)
@@ -1007,7 +1008,7 @@ macro_rules! __msg_send_helper {
         // 1-tuple if there is only one.
         //
         // And use `::<_, _>` for better UI
-        result = $crate::runtime::MessageReceiver::$fn::<_, _>($($fn_args)+, $crate::sel!($($selector)*), ($($argument,)*));
+        result = $crate::__macro_helpers::MsgSend::$fn::<_, _>($($fn_args)+, $crate::sel!($($selector)*), ($($argument,)*));
         result
     });
 }