+++ crates/objc2/src/runtime/mod.rs
@@ -1558,7 +1558,7 @@ mod tests {
 
         let obj = test_utils::custom_object();
         let res: i32 = unsafe {
-            MessageReceiver::send_message(&obj, sel!(test::test::), (1i32, 2i32, 3i32, 4i32))
+            MessageReceiver::send_message(&*obj, sel!(test::test::), (1i32, 2i32, 3i32, 4i32))
         };
         assert_eq!(res, 24);
     }